% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{ensembl_fetch_variant}
\alias{ensembl_fetch_variant}
\title{Fetch Variant Information from Ensembl by rsID}
\usage{
ensembl_fetch_variant(
  rsids,
  endpoint = "http://rest.ensembl.org/variation/homo_sapiens"
)
}
\arguments{
\item{endpoint}{The URL of the Ensembl REST API endpoint. Default is set to the human variation endpoint.}

\item{rsIds}{A character vector of rsIDs to fetch variant information for.}
}
\value{
A data frame containing variant information including start, end, allele strings, ancestral alleles, location, assembly name, sequence region name, and strand.
}
\description{
This function retrieves variant information from the Ensembl REST API using a list of rsIDs.
}
\examples{
\dontrun{
rsIds <- c("rs789012", "rs74733149")
variants <- ensembl_fetch_variant_from_rsid(rsIds)
variants
A tibble: 2 × 10
  seq_region_name strand assembly_name      end ancestral_allele location             coord_system allele_string    start rsid
  <chr>            <int> <chr>            <int> <chr>            <chr>                <chr>        <chr>            <int> <chr>
1 12                   1 GRCh38        28719649 G                12:28719649-28719649 chromosome   G/A           28719649 rs789012
2 1                    1 GRCh38        65609903 A                1:65609903-65609903  chromosome   A/G           65609903 rs74733149
}
}
