% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{ensembl_parse_response}
\alias{ensembl_parse_response}
\title{Parse API Response into a Data Frame}
\usage{
ensembl_parse_response(resp)
}
\arguments{
\item{resp}{A named list where each element is expected to contain a \verb{$mapping} component.}
}
\value{
A \code{data.frame} with one row per item in \code{resp}, including the extracted
\code{mapping} fields and an \code{rsIds} column containing the names of the original list.
}
\description{
Extracts the \code{mapping} element from each object in a response list and
converts the result into a data frame with an added column for the \code{rsIds}.
}
\examples{
\dontrun{
resp <- list(
  rs123 = list(mapping = list(seq_region_name = 1, allele_string = 'G/C/T')),
)
parse_response(resp)
}
}
